clean up of boost to build only the needed libs, clean up of the uintah and nektar++ builds
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@25346 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/svn_bin/bv_support/bv_uintah.sh
@@ -59,7 +59,7 @@ function bv_uintah_initialize_vars
 
 function bv_uintah_info
 {
-export UINTAH_VERSION=${UINTAH_VERSION:-"1.6.0beta"}
+export UINTAH_VERSION=${UINTAH_VERSION:-"1.6.0"}
 export UINTAH_FILE=${UINTAH_FILE:-"Uintah-${UINTAH_VERSION}.tar.gz"}
 export UINTAH_COMPATIBILITY_VERSION=${UINTAH_COMPATIBILITY_VERSION:-"1.6"}
 export UINTAH_BUILD_DIR=${UINTAH_BUILD_DIR:-"Uintah-${UINTAH_VERSION}/optimized"}
@@ -289,7 +289,14 @@ function build_uintah
     #
     info "Installing UINTAH . . ."
 
-    mkdir $VISITDIR/uintah/ || error "Can't make UINTAH install dir."
+    if [[! -e $VISITDIR/uintah ]] ; then
+        mkdir $VISITDIR/uintah/ || error "Can't make UINTAH install dir."
+    fi
+
+    if [[-e $VISITDIR/uintah/$UINTAH_VERSION ]] ; then
+        rm -rf $VISITDIR/uintah/$UINTAH_VERSION || error "Can't remove old UINTAH install dir."
+    fi
+
     mkdir $VISITDIR/uintah/$UINTAH_VERSION/ || error "Can't make UINTAH install dir."
     mkdir $VISITDIR/uintah/$UINTAH_VERSION/$VISITARCH || error "Can't make UINTAH install dir."
     mkdir $VISITDIR/uintah/$UINTAH_VERSION/$VISITARCH/lib || error "Can't make UINTAH install dir."
